Log calls to ExtentManager::recordFor() and getExtent() to stderr.
Dodgy but seems to work.

However, this just indicates getting a pointer to the record/extent,
not necessarily that it is ever actually read/written. eg. as shown
by ExtentManager::extentFor(), which just uses it to simplify the
accounting, not for data access.

To actually check for data read/write, would probably need to use mprotect()
(?), which is likely to be more expensive.

Or, maybe Record and/or Extent can be hooked?